xserver-xorg: Fix the unconditional dependency on virtual/libgbm
Issue: LINCD-11889, LIN1022-2967

For the machines which use the libgbm provided by mesa, since the gbm
is only enabled in mesa when the opengl distro feature is enabled,
adding the dependency on virtual/libgbm for xserver-xorg
unconditionally will break the build of these machines if the opengl
distro feature is not enabled.

There is no such restriction for the machines which use the libgbm
provided by the libmali-xlnx, so only add the virtual/libgbm dependency
for xserver-xorg when the mali400 machine feature is enabled.

(LOCAL REV: NOT UPSTREAM) - send to meta-xilinx on 20230313

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>